oops, broke anemometer reporting (missing cast to float

+++ core1.c
@@ -34,7 +34,7 @@ void GetAnemometerFrequency(void * arg)
     }
     //printf("Transitons = %d\n",tot);
     char ReportStr[50];
-    sprintf(ReportStr,"Anm_freq=%5.1f\n",tot/(num_average));
+    sprintf(ReportStr,"Anm_freq=%5.1f\n",tot/(float)num_average);
     printf("Report: %s",ReportStr);
     if (arg) SendResponse(arg, ReportStr, strlen(ReportStr));
 }
